iOS Plugins: Handle state change notifications, timed mode, bug fixes
- Handle state change notifications and notify the UI (if
  the cause-agent is not the current device).
- Added code to set user attributes when the user creates an account.
  We send device model name, os version and system name (iOS).
- Log into XMPP whenever the app comes to the foreground (prior to this
  we would only login when we needed to send a command).
- Added helper code to retry a request if the initial call to the
  server fails.
- Added code to handle state changes in robot name, schedule state 
  and schedule data change. This completes the notification handling 
  for all robot events.
- We now send a 'schedule_updated' key to the server when the 
  schedule is updated, so that the server can generate state change
  notification for other associated users.

+++ trunk/Smartphone/iOS/SlideiOSPlugin/SlideiOSPlugin/Classes/Helpers/CommandHelper/CommandProtocol.h
@@ -6,6 +6,6 @@
 - (id)getStartRobotCommand;
 - (id)getStopRobotCommand;
 
-- (id)getRobotCommand2WithId:(int)commandId withParams:(NSDictionary *)params andRequestId:(NSString *)requestId;
+- (id)getRobotCommand2WithId:(NSInteger)commandId withParams:(NSDictionary *)params andRequestId:(NSString *)requestId;
 
 @end
