+++ trunk/Smartphone/iOS/SlideiOSPlugin/SlideiOSPlugin/Classes/Helpers/CommandHelper/TCPCommandHelper.m
@@ -33,7 +33,7 @@
 }
 
 
-- (NSData *)getRobotCommand2WithId:(int)commandId withParams:(NSDictionary *)params andRequestId:(NSString *)requestId {
+- (NSData *)getRobotCommand2WithId:(NSInteger)commandId withParams:(NSDictionary *)params andRequestId:(NSString *)requestId {
     debugLog(@"");
     CommandsHelper *commandHelper = [[CommandsHelper alloc] init];
     NSString *command = [NSString stringWithFormat:TEMP_TCP_NEW_ROBOT_COMMAND_FORMAT, [commandHelper versionForCommand], [AppHelper getAppSignature], commandId, requestId, [[NSNumber numberWithDouble:[AppHelper currentTimeStamp]] stringValue], [commandHelper commandRetryCount], [commandHelper commandResponseNeeded], [NeatoUserHelper getNeatoUser].userId, [commandHelper distributionModeForCommandType:@"TCP"], [commandHelper generateXMLForParams:params]];
